requirements.txt
flask==2.3.3
Werkzeug==2.3.7
flask-pymongo==2.3.0
python-dotenv==1.0.0
pymongo==4.5.0